refactor(nem): noexcepts, consts are now variables instead of functions, improve sin and cos

+++ tests/tests_common.hpp
@@ -41,18 +41,18 @@ static void require_valid_basis(nem::float3 n, nem::float3 b1, nem::float3 b2)
     EXPECT_NEAR(nem::dot(b1, b2), 0.f, fEps);
 }
 
-static void range_accuracy_test(float min, float max, int iterations, float (*fun1)(float), float (*fun2)(float), float eps)
+static void range_accuracy_test(double min, double max, int iterations, double (*fun1)(double), double (*fun2)(double), double eps)
 {
-    float length = max - min;
-    float maxerr = 0;
-    float x = min;
-    float step = length / static_cast<float>(iterations);
-    float totalerror = 0;
+    double length = max - min;
+    double maxerr = 0;
+    double x = min;
+    double step = length / static_cast<double>(iterations);
+    double totalerror = 0;
     for (int i = 0; i < iterations; i++)
     {
-        const float nem_result = fun1(x);
-        const float std_result = fun2(x);
-        const float error = nem::abs(nem_result - std_result);
+        const double nem_result = fun1(x);
+        const double std_result = fun2(x);
+        const double error = nem::abs(nem_result - std_result);
         totalerror += error;
         if (error > maxerr)
         {
@@ -61,5 +61,5 @@ static void range_accuracy_test(float min, float max, int iterations, float (*fu
         x += step;
     }
     printf("Max error: |%.6f|\nAverage error: |%.6f|", maxerr, totalerror / iterations);
-    ASSERT_NEAR(maxerr, 0.0f, eps);
+    ASSERT_NEAR(maxerr, 0.0, eps);
 }
